requirements.txt
feedparser==6.0.10
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
tweepy==4.14.0
newsapi==0.1.1
schedule==1.2.0
pytz==2023.12
